+++ src/main/kotlin/net/serverwars/sunsetPlugin/util/player.kt
@@ -3,12 +3,9 @@ package net.serverwars.sunsetPlugin.util
 import net.serverwars.sunsetPlugin.config.Config
 import net.serverwars.sunsetPlugin.domain.server.services.ServerService
 import org.bukkit.Bukkit
-import org.bukkit.OfflinePlayer
 import org.bukkit.entity.Player
 
 fun sendPlayerToMatch(player: Player) {
     val command = "transfer ${ServerService.getServerSlug()}.${Config.getServerwarsMinecraftServerIP()} 25565 ${player.name}"
     Bukkit.getServer().dispatchCommand(Bukkit.getConsoleSender(), command)
 }
-
-fun isKnownPlayer(offlinePlayer: OfflinePlayer): Boolean = offlinePlayer.isOnline || offlinePlayer.hasPlayedBefore()